pyproject.toml
[project]
name = "sar-respond"
version = "0.0.1"
description = "A SAR incident management system to manage CalTopo maps and responders."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
